+++ dash/dash.py
@@ -15,13 +15,14 @@ import logging
 
 from functools import wraps
 
-import plotly
-import dash_renderer
 import flask
 from flask import Flask, Response
 from flask_compress import Compress
 
-from .dependencies import Event, Input, Output, State
+import plotly
+import dash_renderer
+
+from .dependencies import Input, Output, State
 from .resources import Scripts, Css
 from .development.base_component import Component
 from . import exceptions
@@ -29,8 +30,9 @@ from ._utils import AttributeDict as _AttributeDict
 from ._utils import interpolate_str as _interpolate
 from ._utils import format_tag as _format_tag
 from ._utils import generate_hash as _generate_hash
-from . import _watch
 from ._utils import get_asset_path as _get_asset_path
+from ._utils import patch_collections_abc as _patch_collections_abc
+from . import _watch
 from . import _configs
 
 
@@ -76,7 +78,7 @@ class Dash(object):
             name='__main__',
             server=None,
             static_folder='static',
-            assets_folder=None,
+            assets_folder='assets',
             assets_url_path='/assets',
             assets_ignore='',
             include_assets_files=True,
@@ -102,20 +104,14 @@ class Dash(object):
                 See https://github.com/plotly/dash/issues/141 for details.
                 ''', DeprecationWarning)
 
-        name = name if server is None else server.name
-        self._assets_folder = assets_folder or os.path.join(
-            flask.helpers.get_root_path(name), 'assets'
+        self._assets_folder = os.path.join(
+            flask.helpers.get_root_path(name),
+            assets_folder,
         )
         self._assets_url_path = assets_url_path
 
         # allow users to supply their own flask server
         self.server = server or Flask(name, static_folder=static_folder)
-
-        if 'assets' not in self.server.blueprints:
-            self.server.register_blueprint(
-                flask.Blueprint('assets', 'assets',
-                                static_folder=self._assets_folder,
-                                static_url_path=assets_url_path))
 
         env_configs = _configs.env_configs()
 
@@ -146,6 +142,22 @@ class Dash(object):
                 env_configs, 2678400))
         })
 
+        assets_blueprint_name = '{}{}'.format(
+            self.config.routes_pathname_prefix.replace('/', '_'),
+            'dash_assets'
+        )
+
+        self.server.register_blueprint(
+            flask.Blueprint(
+                assets_blueprint_name, name,
+                static_folder=self._assets_folder,
+                static_url_path='{}{}'.format(
+                    self.config.routes_pathname_prefix,
+                    assets_url_path.lstrip('/')
+                )
+            )
+        )
+
         # list of dependencies
         self.callback_map = {}
 
@@ -159,10 +171,9 @@ class Dash(object):
             Compress(self.server)
 
         @self.server.errorhandler(exceptions.PreventUpdate)
-        def _handle_error(error):
+        def _handle_error(_):
             """Handle a halted callback and return an empty 204 response"""
-            print(error, file=sys.stderr)
-            return ('', 204)
+            return '', 204
 
         # static files from the packages
         self.css = Css()
@@ -265,7 +276,7 @@ class Dash(object):
         return self._layout
 
     def _layout_value(self):
-        if isinstance(self._layout, collections.Callable):
+        if isinstance(self._layout, _patch_collections_abc('Callable')):
             self._cached_layout = self._layout()
         else:
             self._cached_layout = self._layout
@@ -274,8 +285,8 @@ class Dash(object):
     @layout.setter
     def layout(self, value):
         if (not isinstance(value, Component) and
-                not isinstance(value, collections.Callable)):
-            raise Exception(
+                not isinstance(value, _patch_collections_abc('Callable'))):
+            raise exceptions.NoLayoutException(
                 ''
                 'Layout must be a dash component '
                 'or a function that returns '
@@ -301,7 +312,7 @@ class Dash(object):
         )
         missing = [missing for check, missing in checks if not check]
         if missing:
-            raise Exception(
+            raise exceptions.InvalidIndexException(
                 'Did you forget to include {} in your index string ?'.format(
                     ', '.join('{%' + x + '%}' for x in missing)
                 )
@@ -391,11 +402,12 @@ class Dash(object):
                             namespace=resource['namespace']
                         ))
             elif 'external_url' in resource:
-                if isinstance(resource['external_url'], str):
-                    srcs.append(resource['external_url'])
-                else:
-                    for url in resource['external_url']:
-                        srcs.append(url)
+                if not is_dynamic_resource:
+                    if isinstance(resource['external_url'], str):
+                        srcs.append(resource['external_url'])
+                    else:
+                        for url in resource['external_url']:
+                            srcs.append(url)
             elif 'absolute_path' in resource:
                 raise Exception(
                     'Serving files from absolute_path isn\'t supported yet'
@@ -475,14 +487,14 @@ class Dash(object):
     # Serve the JS bundles for each package
     def serve_component_suites(self, package_name, path_in_package_dist):
         if package_name not in self.registered_paths:
-            raise exceptions.InvalidResourceError(
+            raise exceptions.DependencyException(
                 'Error loading dependency.\n'
                 '"{}" is not a registered library.\n'
                 'Registered libraries are: {}'
                 .format(package_name, list(self.registered_paths.keys())))
 
         elif path_in_package_dist not in self.registered_paths[package_name]:
-            raise exceptions.InvalidResourceError(
+            raise exceptions.DependencyException(
                 '"{}" is registered but the path requested is not valid.\n'
                 'The path requested: "{}"\n'
                 'List of registered paths: {}'
@@ -546,7 +558,7 @@ class Dash(object):
 
         if missing:
             plural = 's' if len(missing) > 1 else ''
-            raise Exception(
+            raise exceptions.InvalidIndexException(
                 'Missing element{pl} {ids} in index.'.format(
                     ids=', '.join(missing),
                     pl=plural
@@ -612,7 +624,6 @@ class Dash(object):
                 },
                 'inputs': v['inputs'],
                 'state': v['state'],
-                'events': v['events']
             } for k, v in self.callback_map.items()
         ])
 
@@ -623,7 +634,7 @@ class Dash(object):
             'Use `callback` instead. `callback` has a new syntax too, '
             'so make sure to call `help(app.callback)` to learn more.')
 
-    def _validate_callback(self, output, inputs, state, events):
+    def _validate_callback(self, output, inputs, state):
         # pylint: disable=too-many-branches
         layout = self._cached_layout or self._layout_value()
 
@@ -642,8 +653,7 @@ class Dash(object):
 
         for args, obj, name in [([output], Output, 'Output'),
                                 (inputs, Input, 'Input'),
-                                (state, State, 'State'),
-                                (events, Event, 'Event')]:
+                                (state, State, 'State')]:
 
             if not isinstance(args, list):
                 raise exceptions.IncorrectTypeException(
@@ -664,7 +674,7 @@ class Dash(object):
                                           'supress_callback_exceptions') and
                         arg.component_id not in layout and
                         arg.component_id != getattr(layout, 'id', None)):
-                    raise exceptions.NonExistantIdException('''
+                    raise exceptions.NonExistentIdException('''
                         Attempting to assign a callback to the
                         component with the id "{}" but no
                         components with id "{}" exist in the
@@ -697,7 +707,7 @@ class Dash(object):
                             component.available_properties and not
                             any(arg.component_property.startswith(w) for w in
                                 component.available_wildcard_properties)):
-                        raise exceptions.NonExistantPropException('''
+                        raise exceptions.NonExistentPropException('''
                             Attempting to assign a callback with
                             the property "{}" but the component
                             "{}" doesn't have "{}" as a property.\n
@@ -711,32 +721,20 @@ class Dash(object):
                             component.available_properties).replace(
                                 '    ', ''))
 
-                    if (hasattr(arg, 'component_event') and
-                            arg.component_event not in
-                            component.available_events):
-                        raise exceptions.NonExistantEventException('''
-                            Attempting to assign a callback with
-                            the event "{}" but the component
-                            "{}" doesn't have "{}" as an event.\n
-                            Here is a list of the available events in "{}":
-                            {}
-                        '''.format(
-                            arg.component_event,
-                            arg.component_id,
-                            arg.component_event,
-                            arg.component_id,
-                            component.available_events).replace('    ', ''))
+                    if hasattr(arg, 'component_event'):
+                        raise exceptions.NonExistentEventException('''
+                            Events have been removed.
+                            Use the associated property instead.
+                        ''')
 
-        if state and not events and not inputs:
-            raise exceptions.MissingEventsException('''
+        if state and not inputs:
+            raise exceptions.MissingInputsException('''
                 This callback has {} `State` {}
-                but no `Input` elements or `Event` elements.\n
-                Without `Input` or `Event` elements, this callback
+                but no `Input` elements.\n
+                Without `Input` elements, this callback
                 will never get called.\n
                 (Subscribing to input components will cause the
-                callback to be called whenever their values
-                change and subscribing to an event will cause the
-                callback to be called whenever the event is fired.)
+                callback to be called whenever their values change.)
             '''.format(
                 len(state),
                 'elements' if len(state) > 1 else 'element'
@@ -878,8 +876,8 @@ class Dash(object):
     # TODO - Check this map for recursive or other ill-defined non-tree
     # relationships
     # pylint: disable=dangerous-default-value
-    def callback(self, output, inputs=[], state=[], events=[]):
-        self._validate_callback(output, inputs, state, events)
+    def callback(self, output, inputs=[], state=[]):
+        self._validate_callback(output, inputs, state)
 
         callback_id = '{}.{}'.format(
             output.component_id, output.component_property
@@ -892,10 +890,6 @@ class Dash(object):
             'state': [
                 {'id': c.component_id, 'property': c.component_property}
                 for c in state
-            ],
-            'events': [
-                {'id': c.component_id, 'event': c.component_event}
-                for c in events
             ]
         }
 
@@ -1055,7 +1049,6 @@ class Dash(object):
     def get_asset_url(self, path):
         asset = _get_asset_path(
             self.config.requests_pathname_prefix,
-            self.config.routes_pathname_prefix,
             path,
             self._assets_url_path.lstrip('/')
         )