+++ dash/exceptions.py
@@ -10,15 +10,15 @@ class CallbackException(DashException):
     pass
 
 
-class NonExistantIdException(CallbackException):
+class NonExistentIdException(CallbackException):
     pass
 
 
-class NonExistantPropException(CallbackException):
+class NonExistentPropException(CallbackException):
     pass
 
 
-class NonExistantEventException(CallbackException):
+class NonExistentEventException(CallbackException):
     pass
 
 
@@ -30,7 +30,7 @@ class IncorrectTypeException(CallbackException):
     pass
 
 
-class MissingEventsException(CallbackException):
+class MissingInputsException(CallbackException):
     pass
 
 
@@ -63,4 +63,16 @@ class InvalidConfig(DashException):
 
 
 class InvalidResourceError(DashException):
+    pass
+
+
+class InvalidIndexException(DashException):
+    pass
+
+
+class DependencyException(DashException):
+    pass
+
+
+class ResourceException(DashException):
     pass