+++ tests/development/test_base_component.py
@@ -7,11 +7,9 @@ import shutil
 import unittest
 import plotly
 
-from dash.development.base_component import (
-    Component,
-    _explicitize_args)
+from dash.development.base_component import Component
 from dash.development._py_components_generation import generate_class_string, generate_class_file, generate_class, \
-    create_docstring, parse_events, js_to_py_type
+    create_docstring, prohibit_events, js_to_py_type
 
 Component._prop_names = ('id', 'a', 'children', 'style', )
 Component._type = 'TestComponent'
@@ -282,7 +280,7 @@ class TestComponent(unittest.TestCase):
 
         c3b = Component(id='3')
         self.assertEqual(c5['3'], c3)
-        self.assertTrue(c5['3'] is not '3')
+        self.assertTrue(c5['3'] != '3')
         self.assertTrue(c5['3'] is not c3b)
 
         c5['3'] = c3b
@@ -547,17 +545,23 @@ class TestGenerateClassFile(unittest.TestCase):
     def tearDown(self):
         shutil.rmtree('TableComponents')
 
+    def assert_no_trailing_spaces(self, s):
+        for line in s.split('\n'):
+            self.assertEqual(line, line.rstrip())
+
     def test_class_string(self):
         self.assertEqual(
             self.expected_class_string,
             self.component_class_string
         )
+        self.assert_no_trailing_spaces(self.component_class_string)
 
     def test_class_file(self):
         self.assertEqual(
             self.expected_class_string,
             self.written_class_string
         )
+        self.assert_no_trailing_spaces(self.written_class_string)
 
 
 class TestGenerateClass(unittest.TestCase):
@@ -685,10 +689,10 @@ class TestGenerateClass(unittest.TestCase):
     def test_docstring(self):
         assert_docstring(self.assertEqual, self.ComponentClass.__doc__)
 
-    def test_events(self):
+    def test_no_events(self):
         self.assertEqual(
-            self.ComponentClass().available_events,
-            ['restyle', 'relayout', 'click']
+            hasattr(self.ComponentClass(), 'available_events'),
+            False
         )
 
     # This one is kind of pointless now
@@ -793,11 +797,11 @@ class TestMetaDataConversions(unittest.TestCase):
             ['optionalObjectWithShapeAndNestedDescription', '\n'.join([
 
                 "dict containing keys 'color', 'fontSize', 'figure'.",
-                "Those keys have the following types: ",
+                "Those keys have the following types:",
                 "  - color (string; optional)",
                 "  - fontSize (number; optional)",
                 "  - figure (optional): Figure is a plotly graph object. figure has the following type: dict containing keys 'data', 'layout'.",  # noqa: E501
-                "Those keys have the following types: ",
+                "Those keys have the following types:",
                 "  - data (list; optional): data is a collection of traces",
                 "  - layout (dict; optional): layout describes the rest of the figure"  # noqa: E501
 
@@ -815,18 +819,16 @@ class TestMetaDataConversions(unittest.TestCase):
 
             ['in', 'string'],
 
-            ['id', 'string'],
-
-            ['dashEvents', "a value equal to: 'restyle', 'relayout', 'click'"]
+            ['id', 'string']
         ])
 
     def test_docstring(self):
         docstring = create_docstring(
             'Table',
             self.data['props'],
-            parse_events(self.data['props']),
             self.data['description'],
         )
+        prohibit_events(self.data['props']),
         assert_docstring(self.assertEqual, docstring)
 
     def test_docgen_to_python_args(self):
@@ -871,7 +873,7 @@ def assert_docstring(assertEqual, docstring):
             "following type: dict containing keys "
             "'color', 'fontSize', 'figure'.",
 
-            "Those keys have the following types: ",
+            "Those keys have the following types:",
             "  - color (string; optional)",
             "  - fontSize (number; optional)",
 
@@ -879,7 +881,7 @@ def assert_docstring(assertEqual, docstring):
             "figure has the following type: dict containing "
             "keys 'data', 'layout'.",
 
-            "Those keys have the following types: ",
+            "Those keys have the following types:",
             "  - data (list; optional): data is a collection of traces",
 
             "  - layout (dict; optional): layout describes "
@@ -894,8 +896,6 @@ def assert_docstring(assertEqual, docstring):
             '- aria-* (string; optional)',
             '- in (string; optional)',
             '- id (string; optional)',
-            '',
-            "Available events: 'restyle', 'relayout', 'click'",
             '        '
             ])[i]
                    )
@@ -931,7 +931,7 @@ class TestFlowMetaDataConversions(unittest.TestCase):
             ['optionalSignature(shape)', '\n'.join([
 
                 "dict containing keys 'checked', 'children', 'customData', 'disabled', 'label', 'primaryText', 'secondaryText', 'style', 'value'.",
-                "Those keys have the following types: ",
+                "Those keys have the following types:",
                 "- checked (boolean; optional)",
                 "- children (a list of or a singular dash component, string or number; optional)",
                 "- customData (bool | number | str | dict | list; required): A test description",
@@ -947,9 +947,9 @@ class TestFlowMetaDataConversions(unittest.TestCase):
             ['requiredNested', '\n'.join([
 
                 "dict containing keys 'customData', 'value'.",
-                "Those keys have the following types: ",
+                "Those keys have the following types:",
                 "- customData (required): . customData has the following type: dict containing keys 'checked', 'children', 'customData', 'disabled', 'label', 'primaryText', 'secondaryText', 'style', 'value'.",
-                "  Those keys have the following types: ",
+                "  Those keys have the following types:",
                 "  - checked (boolean; optional)",
                 "  - children (a list of or a singular dash component, string or number; optional)",
                 "  - customData (bool | number | str | dict | list; required)",
@@ -968,9 +968,9 @@ class TestFlowMetaDataConversions(unittest.TestCase):
         docstring = create_docstring(
             'Flow_component',
             self.data['props'],
-            parse_events(self.data['props']),
             self.data['description'],
         )
+        prohibit_events(self.data['props']),
         assert_flow_docstring(self.assertEqual, docstring)
 
     def test_docgen_to_python_args(self):
@@ -1010,7 +1010,7 @@ def assert_flow_docstring(assertEqual, docstring):
             "'children', 'customData', 'disabled', 'label', 'primaryText', 'secondaryText', "
             "'style', 'value'.",
 
-            "  Those keys have the following types: ",
+            "  Those keys have the following types:",
             "  - checked (boolean; optional)",
             "  - children (a list of or a singular dash component, string or number; optional)",
             "  - customData (bool | number | str | dict | list; required): A test description",
@@ -1024,13 +1024,13 @@ def assert_flow_docstring(assertEqual, docstring):
             "- requiredNested (required): . requiredNested has the following type: dict containing "
             "keys 'customData', 'value'.",
 
-            "  Those keys have the following types: ",
+            "  Those keys have the following types:",
 
             "  - customData (required): . customData has the following type: dict containing "
             "keys 'checked', 'children', 'customData', 'disabled', 'label', 'primaryText', "
             "'secondaryText', 'style', 'value'.",
 
-            "    Those keys have the following types: ",
+            "    Those keys have the following types:",
             "    - checked (boolean; optional)",
             "    - children (a list of or a singular dash component, string or number; optional)",
             "    - customData (bool | number | str | dict | list; required)",
@@ -1041,7 +1041,5 @@ def assert_flow_docstring(assertEqual, docstring):
             "    - style (dict; optional)",
             "    - value (bool | number | str | dict | list; required)",
             "  - value (bool | number | str | dict | list; required)",
-            "",
-            "Available events: "
         ])[i]
                    )